+++ database/migrations/2023_05_18_161405_create_comprovante_despesas_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->increments('id');
             $table->string('anexoDespesa')->nullable();
             $table->string('descricao')->nullable();
-            $table->integer('valorReais')->nullable();
-            $table->integer('valorDolar')->nullable();
+            $table->float('valorReais')->nullable();
+            $table->float('valorDolar')->nullable();
             $table->dateTime('dataOcorrencia');
 
             //Referencia a tabela de AV
